separate tests in pkg/receiver
