+++ components/event-publisher-proxy/pkg/receiver/receiver_test.go
@@ -1,4 +1,7 @@
-package receiver
+//go:build unit
+// +build unit
+
+package receiver_test
 
 import (
 	"context"
@@ -9,7 +12,7 @@ import (
 
 	"github.com/kyma-project/kyma/components/eventing-controller/logger"
 
-	testingutils "github.com/kyma-project/kyma/components/event-publisher-proxy/testing"
+	sut "github.com/kyma-project/kyma/components/event-publisher-proxy/pkg/receiver"
 )
 
 // a mocked http.Handler
@@ -18,17 +21,6 @@ type testHandler struct{}
 func (h *testHandler) ServeHTTP(http.ResponseWriter, *http.Request) {}
 
 var _ http.Handler = (*testHandler)(nil)
-
-func TestNewHttpMessageReceiver(t *testing.T) {
-	port := testingutils.GeneratePortOrDie()
-	r := NewHTTPMessageReceiver(port)
-	if r == nil {
-		t.Fatalf("Could not create HTTPMessageReceiver")
-	}
-	if r.port != port {
-		t.Errorf("Port should be: %d is: %d", port, r.port)
-	}
-}
 
 // Test that the receiver shutdown when receiving stop signal
 func TestStartListener(t *testing.T) {
@@ -78,6 +70,6 @@ func TestStartListener(t *testing.T) {
 	}
 }
 
-func fixtureReceiver() *HTTPMessageReceiver {
-	return NewHTTPMessageReceiver(0)
+func fixtureReceiver() *sut.HTTPMessageReceiver {
+	return sut.NewHTTPMessageReceiver(0)
 }
